# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/exam/all-exams": {
      "get": {
        "operationId": "ExamPermissionController_getListExam",
        "summary": "Get list exam for user",
        "parameters": [
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isFinished",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "examId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "score",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "examType",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classRoomName",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/all-practice-exams/{teacherId}": {
      "get": {
        "operationId": "ExamPermissionController_getListPracticeExam",
        "summary": "Get list practice exam for teacher to score",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "createdDate",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isFinished",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "examId",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "score",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "examType",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classRoomName",
            "required": false,
            "in": "query",
            "description": "not required",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/exam-saved": {
      "post": {
        "operationId": "ExamPermissionController_saveExamsSaved",
        "summary": "Save exam answer for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EXAM"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/submit-practice": {
      "post": {
        "operationId": "ExamPermissionController_submitPracticeTest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeExamScoringDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/exam-scoring": {
      "post": {
        "operationId": "ExamPermissionController_examScoring",
        "summary": "Scoring exam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamScoringDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/practice-exam-scoring": {
      "post": {
        "operationId": "ExamPermissionController_practiceExamScoring",
        "summary": "Scoring practice exam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeExamScoringDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/reset/{examId}": {
      "post": {
        "operationId": "ExamPermissionController_resetExam",
        "summary": "Reset exam for redo",
        "parameters": [
          {
            "name": "examId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/practice-exam/reset/{examId}": {
      "post": {
        "operationId": "ExamPermissionController_resetPracticeExam",
        "summary": "Reset practice exam for redo",
        "parameters": [
          {
            "name": "examId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeExamAttempt"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/practice-exams/{examId}": {
      "get": {
        "operationId": "ExamPermissionController_getDetailPracticeExam",
        "summary": "Get detail practice exam",
        "parameters": [
          {
            "name": "examId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamVocabulary"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/detail-exam/{id}": {
      "get": {
        "operationId": "ExamPermissionController_getDetailExam",
        "summary": "Get detail exam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamQuestion"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/practice-exams-to-score/{examId}/{userId}": {
      "get": {
        "operationId": "ExamPermissionController_getDetailPracticeExamToScore",
        "summary": "Get detail practice exam to score",
        "parameters": [
          {
            "name": "examId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamVideo"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/practice-exams": {
      "post": {
        "operationId": "ExamPermissionController_addPracticeExam",
        "summary": "Create practice exam ",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePracticeExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamB"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/add-exam": {
      "post": {
        "operationId": "ExamPermissionController_addExam",
        "summary": "Create exam ",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamB"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
    "/api/exam/edit-exam": {
      "put": {
        "operationId": "ExamPermissionController_editExam",
        "summary": "Edit exam ",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamB"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bluequant API APIs document",
    "description": "",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "DefaultResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "required"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDataResponse"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "ExamAttempt": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "bigint"
          },
          "examId": {
            "type": "bigint"
          },
          "score": {
            "type": "string"
          },
          "isFinished": {
            "type": "bit",
            "default": 0
          }
        },
        "required": [
          "studentId",
          "examId",
          "score",
          "isFinished"
        ]
      },
...
      "EXAM": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "updatedAt": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "name": {
            "type": "string"
          },
          "classRoomId": {
            "type": "bigint"
          },
          "createdBy": {
            "type": "string"
          },
          "private": {
            "type": "bit"
          }
        },
        "required": [
          "createdDate",
          "updatedAt",
          "name",
          "classRoomId",
          "createdBy",
          "private"
        ]
      },
      "PracticeExamScoringDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "not required"
          },
          "examId": {
            "type": "number",
            "description": "not required"
          },
          "score": {
            "type": "number",
            "description": "not required"
          },
          "isFinished": {
            "type": "boolean",
            "description": "not required"
          },
          "attemptCount": {
            "type": "number",
            "description": "not required"
          }
        }
      },
      "ExamScoringDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "not required"
          },
          "examId": {
            "type": "number",
            "description": "not required"
          },
          "score": {
            "type": "number",
            "description": "not required"
          },
          "isFinished": {
            "type": "boolean",
            "description": "not required"
          },
          "attemptCount": {
            "type": "number",
            "description": "not required"
          },
          "type": {
            "type": "string",
            "description": "not required"
          }
        }
      },
      "PracticeExamAttempt": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "studentId": {
            "type": "bigint"
          },
          "examId": {
            "type": "bigint"
          },
          "score": {
            "type": "string"
          },
          "isFinished": {
            "type": "bit",
            "default": 0
          }
        },
        "required": [
          "createdDate",
          "studentId",
          "examId",
          "score",
          "isFinished"
        ]
      },
      "ResetExamDto": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "number",
            "description": "not required"
          },
          "userId": {
            "type": "number",
            "description": "not required"
          }
        }
      },
      "ExamVocabulary": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "vocabularyExamId": {
            "type": "number",
            "description": "required",
            "default": 1
          },
          "vocabularyId": {
            "type": "bigint"
          },
          "examId": {
            "type": "bigint"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "createdDate",
          "vocabularyExamId",
          "vocabularyId",
          "examId",
          "content"
        ]
      },
      "ExamQuestion": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "bigint"
          },
          "examId": {
            "type": "bigint"
          }
        },
        "required": [
          "questionId",
          "examId"
        ]
      },
      "ExamVideo": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "videoExamId": {
            "type": "number",
            "description": "required",
            "default": 1
          },
          "videoUrl": {
            "type": "string"
          },
          "aiAnswer": {
            "type": "string"
          },
          "examId": {
            "type": "bigint"
          },
          "userId": {
            "type": "bigint"
          }
        },
        "required": [
          "createdDate",
          "videoExamId",
          "videoUrl",
          "aiAnswer",
          "examId",
          "userId"
        ]
      },
...
      "CreatePracticeExamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "required",
            "default": "name"
          },
          "classRoomId": {
            "type": "number",
            "description": "required"
          },
          "isPrivate": {
            "type": "boolean",
            "description": "not required"
          },
          "practiceWords": {
            "description": "required",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeWordDto"
            }
          }
        },
        "required": [
          "name",
          "classRoomId",
          "practiceWords"
        ]
      },
...
    }
  }
}